Реальзовать API Private: События

+++ ewm-main-service/src/main/java/ru/practicum/dto/event/EventShortDto.java
@@ -24,5 +24,5 @@ public class EventShortDto {
     private UserShortDto initiator;
     private boolean paid;
     private String title;
-    private int views;
+    private long views;
 }
