+++ ewm-main-service/src/main/java/ru/practicum/repository/CategoriesRepository.java
@@ -7,7 +7,7 @@ import ru.practicum.model.Category;
 
 import java.util.List;
 
-public interface CategoriesRepository extends CrudRepository<Category, Integer> {
+public interface CategoriesRepository extends CrudRepository<Category, Long> {
     @Query("SELECT c FROM Category c ORDER BY c.id LIMIT :size OFFSET :from")
     List<Category> findAllLimit(@Param("from") int from, @Param("size") int size);
 }
